Validation

+++ src/main/java/sda/spring/mvc/controller/UserController.java
@@ -4,10 +4,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import sda.spring.mvc.model.User;
 import sda.spring.mvc.model.dto.UserDTO;
 import sda.spring.mvc.service.UserService;
+import sda.spring.mvc.service.exception.UserNotFoundException;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -23,8 +27,9 @@ import java.util.List;
  */
 @Controller
 public class UserController {
-    private UserService userService;
-    private final String index = "index";
+    private final UserService userService;
+    private static final String INDEX = "index";
+    private static final String USERS = "users";
 
     @Autowired
     private UserController(UserService userService) {
@@ -34,31 +39,38 @@ public class UserController {
     @GetMapping("/")
     public String listAll(Model model) {
         List<UserDTO> userDTOS = userService.findAll();
-        model.addAttribute("users", userDTOS);
-        return index;
+        model.addAttribute(USERS, userDTOS);
+        return INDEX;
     }
 
+    //validarea campurilor prin DTO
+    //@ModelAttribute("user") face legatura dintre obiectul meu (DTO) si ceea ce returneaza formularul
+    //de submit din HTML
     @PostMapping("/adduser")
-    public String save(@Valid User user, BindingResult bindingResult, Model model) {
+    public String save(@Valid @ModelAttribute("user") UserDTO userDTO, BindingResult bindingResult, Model model) {
         //daca rezultatul are erori intorc aceeasi pagina
         if (bindingResult.hasErrors()) {
             return "user-add";
         }
-        userService.save(user);
-        model.addAttribute("users", userService.findAll());
-        return index;
+        userService.save(userDTO);
+        model.addAttribute(USERS, userService.findAll());
+        return INDEX;
     }
 
     @GetMapping("/signup")
-    public String getSignupPage(User user) {
+    public String getSignupPage(@ModelAttribute("user") User user) {
         return "user-add";
     }
 
     @GetMapping("/delete/{id}")
     public String deleteUser(Model model, @PathVariable("id") Long id) {
+        UserDTO userDTO = userService.getById(id);
+        if (userDTO == null) {
+            throw new UserNotFoundException();
+        }
         userService.remove(id);
-        model.addAttribute("users", userService.findAll());
-        return index;
+        model.addAttribute(USERS, userService.findAll());
+        return INDEX;
     }
 
     @GetMapping("/edit/{id}")
@@ -69,9 +81,12 @@ public class UserController {
     }
 
     @PostMapping("/update/{id}")
-    public String update(@Valid User user, @PathVariable Long id, BindingResult result, Model model) {
-        userService.save(user);
-        model.addAttribute("users", userService.findAll());
+    public String update(@Valid @ModelAttribute("user") UserDTO userDTO, BindingResult result, Model model) {
+        if (result.hasErrors()) {
+            return "user-edit";
+        }
+        userService.save(userDTO);
+        model.addAttribute(USERS, userService.findAll());
         return "redirect:/";
     }
 }
